OK LES GO?????

+++ online_bulletin.rpy
@@ -746,8 +746,9 @@ label splashscreen:
     $ update_news()
 
 label quit:
-    $ fetch_rpy(persistent.current_news["bullet_file"])
-    $ delete_file()
+    if persistent.current_news != None:
+        $ fetch_rpy(persistent.current_news["bullet_file"])
+        $ delete_file()
     $ persistent.news = None
     $ persistent.current_news = None
     $ cleanup_temp_files()
